drm/armada: add mmp2 support
This pretty much boils down to setting the LCD_CFG_SCLK_DIV register, and
is tailored to the OLPC XO-1.75. I have no idea what are the meanings of
most bits there, so I'm just making sure it ends up being 0x40001102.

This means that the selection of the source clock is hardwired. Apparently
the bit 30 selects the AXI bus clock as base clock for the pixel clock.
It is not known to me what other options are there.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>

---
Notes; perhaps James or someone else with a datasheet could help me with
this.

I'm somewhat confused about the clock selection. The firmware contains
this line:

  h# 00000700 value pmua-disp-clk-sel  \ PLL1 / 7 -> 113.86 MHz

However, the Linux clock driver seems to consider the value of 7 to
configure the divisor to 8, so the resulting clock would end up being
100 MHz. Also, the clk-of-mmp2 driver suggests PLL1 outputs 800 MHz,
dividing that by 7 would end up being 114.29 MHz, not 113.86 MHz.

If the same logic was used here as Armada 510 driver uses, we'll end up
with the divisor of 1 in LCD_CFG_SCLK_DIV and disp0_div divisor of 16.
Would that be good also? (Do perhaps any of the bits in LCD_CFG_SCLK_DIV
allow for a fractional divisor, allowing us to get the clock we need
more precisely?)

The pxa168fb driver as used on the stock OLPC software just avoids
touching the register, preserving the value set from the firmware.