clk: mmp2: create a power domain for the GPU core
The power management unit on MMP2 is able to gate clock for the GC860 GPU.

There's some special dance required to initialize the unit after the power
has been enabled. If not followed, either the GPU's memory interface or the
GPU core doesn't work or the SoC just hangs.

Once the power has been applied to the GPU block, it doesn't seem
possible to turn it off entirely and initialize again.

As the data sheet is missing, neither the details about initialization, nor
the reason why it can't be reinitialized are understood. The meaning of
most bits in the APMU_GPU register are partially described in [1].

[1] http://lists.laptop.org/pipermail/devel/2019-April/039053.html

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>