clk: mmp2: Create a domain for the GPU power island
The GC860 GPU on MMP2 and GC2000 & GC300 pair on MMP3 are on a separate
power island which needs to be enabled before the GPUs can be clocked and
used.  After it's enabled, some clock & reset dance is needed for the GPUs
to initialize themselves, otherwise they hang or misbehave.

On MMP2, it seems impossible to re-initialize the GPUs once they've been
brought up. Perhaps that's a silicon bug, because on MMP3 just casually
asserting and releasing the resets does the job.

There's no documentation, which is why this might not be the most
optimal thing to do. The MMP2 power on sequence comes from [1],
the MMP3 specific bits come from Marvell's tree.

Tested on a OLPC XO 1.75 laptop (MMP2) and Dell Wyse 3020 (MMP3).

[1] http://lists.laptop.org/pipermail/devel/2019-April/039053.html

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>