Drivers: hv: Move Hyper-V clocksource code to new clocksource driver
Code for the Hyper-V specific clocksources is currently mixed
in with other Hyper-V code. Move the code to the Hyper-V specific
driver in the "clocksource" directory, while separating out
ISA dependencies so that the clocksource driver remains ISA
independent. Update the Hyper-V initialization code to call
initialization and cleanup routines since the Hyper-V synthetic
timers are not independently enumerated in ACPI. Update Hyper-V
clocksource users KVM and VDSO to get definitions from a new
include file.

No behavior is changed and no new functionality is added.

Suggested-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Michael Kelley <mikelley@microsoft.com>
Signed-off-by: Sasha Levin <sashal@kernel.org>