mm/memremap: rename and consolidate SECTION_SIZE
Trying to activate ZONE_DEVICE for arm64 reveals that memremap's internal
helpers for sparsemem sections conflict with arm64's definitions for
hugepages, which inherit the name of "sections" from earlier versions of
the ARM architecture.

Disambiguate memremap (and now HMM too) by propagating sparsemem's PA_
prefix, to clarify that these values are in terms of addresses rather than
PFNs (and because it's a heck of a lot easier than changing all the arch
code).  SECTION_MASK is unused, so it can just go.

[anshuman.khandual@arm.com: consolidate mm/hmm.c instance and updated the commit message]
Link: http://lkml.kernel.org/r/d291c21d1b401e324f0e0bf23e1b3fdb4159d425.1558547956.git.robin.murphy@arm.com
Signed-off-by: Robin Murphy <robin.murphy@arm.com>
Signed-off-by: Anshuman Khandual <anshuman.khandual@arm.com>
Acked-by: Michal Hocko <mhocko@suse.com>
Reviewed-by: David Hildenbrand <david@redhat.com>
Acked-by: Dan Williams <dan.j.williams@intel.com>
Cc: Will Deacon <will.deacon@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: Ira Weiny <ira.weiny@intel.com>
Cc: Jerome Glisse <jglisse@redhat.com>
Cc: Michael Ellerman <mpe@ellerman.id.au>
Cc: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>