dt-bindings: display: armada: Convert to json-schema
This is a rather straightforward conversion. Compared to the original
document this adds documentation ofr the digital vide port node and
fixes the clock name in the example.

The file is also renamed to something that better reflects what the
binding describes (there will be more hardware than just Dove supported)
and moved from a separate directory.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>

---
Changes since v4:
- Replaced the following patches from previous version of the set with
  this one:
  dt-bindings: display: armada: Rename the binding doc file
  dt-bindings: display: armada: Improve the LCDC documentation