drm/gud: Add functionality for the USB gadget side
Since the USB gadget/device has to reach into the DRM internals, part of
the code is placed in the DRM subsystem as a separate module. All calls
into this module runs in process context and are serialized, except one
function that runs in interrupt context.

Since both the gadget side and the DRM side can disable the gadget,
special care has been taken to make that safe.

v2:
- Use drm_client_modeset_disable()

Signed-off-by: Noralf Trønnes <noralf@tronnes.org>