ARM: dts: mmp3-dell-ariel: Add the power button node
This adds support for the power button attached to the Embedded Controller
on a Dell Wyse 3020 "Ariel" board.

However, while the EC itself is controlled via I2C, the input capability
for the power button acts as a separate device attached to the SPI, hence
it has a separate device node.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>