docs: arm: marvell: replace stale links with archive links
Marvell has an annoying habit of moving stuff around their web site
every full moon, more of ten than not removing documents altogether.

At this point basically none but four of the links still works and even
those that work today weren't working for a long period of time
previously. That is a shame, because short of the product briefs, the
documents tend to be useful.

Let's replace them with known working versions of IA's Wayback Machine
links. That seems to be about the only way of getting a URL that's going
to work the next week.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>