ARM: dts: mmp3-olpc-xo-4: Add a DTS file for OLPC XO-4 laptop
The MMP3-based OLPC XO-4 now boots with the mainline kernel. Let's add a
DTS file.

This is not absolutely needed to boot the machine because uses Open
Firmware that provides its own device tree. However, having an in-tree DTS
is still useful in several ways: it provides a convenient way of checking
how the drivers are actually used without possessing the machine, makes
it possible to use the validation tools against the bindings, etc.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>