lib/hexdump: introduce DUMP_PREFIX_UNHASHED for unhashed addresses
Patch series "introduce DUMP_PREFIX_UNHASHED for hex dumps".

First patch updates print_hex_dump() and related functions to
allow callers to print hex dumps with unhashed addresses.  It
adds a new prefix type, so existing code is unchanged.

Second patch changes a page poising function to use the new address type.
This is an example of a change.

Hashed addresses make very little sense for hex dumps, which print
addresses in 16- or 32-byte increments.  Typical use-case is to correlate
an addresses in between one of these increments with some other address,
but that can't be done if the addresses are hashed.  I expect most
developers to want to replace their usage of DUMP_PREFIX_ADDRESS with
DUMP_PREFIX_UNHASHED, now that they have the opportunity.

This patch (of 2):

Hashed addresses are useless in hexdumps unless you're comparing with
other hashed addresses, which is unlikely.  However, there's no need to
break existing code, so introduce a new prefix type that prints unhashed
addresses.

Link: https://lkml.kernel.org/r/20210116220950.47078-1-timur@kernel.org
Link: https://lkml.kernel.org/r/20210116220950.47078-2-timur@kernel.org
Signed-off-by: Timur Tabi <timur@kernel.org>
Reviewed-by: Petr Mladek <pmladek@suse.com>
Cc: Roman Fietze <roman.fietze@magna.com>
Cc: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>
Cc: Kees Cook <keescook@chromium.org>
Cc: Steven Rostedt <rostedt@goodmis.org>
Cc: John Ogness <john.ogness@linutronix.de>
Cc: Akinobu Mita <akinobu.mita@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>